fix: limit changed to DecimalField

+++ categories/models.py
@@ -24,7 +24,7 @@ class Category(models.Model):
     name = models.CharField(
         max_length=20, choices=CategoryName.choices, default=CategoryName.other
     )
-    limit = models.FloatField(default=0)
+    limit = models.DecimalField(max_digits=1000, decimal_places=2, default=0)
 
     user = models.ForeignKey(
         "users.User",
